upload task 10

+++ 101.OOP/Abstr&Interface/10.Military Elite2/Factory/SoldierFactory.cs
@@ -31,7 +31,7 @@ namespace Military_Elite2.Factory
                     return new Commando(id, firsName, lastName, salary, corps);
 
                 case SoldierType.Spy:
-                    return new Spy(id, firsName, lastName, (int)salary);
+                    return new Spy(id, firsName, lastName, salary);
 
                 default:
                     return null;
