+++ 101.OOP/Abstr&Interface/10.Military Elite2/Models/Spy.cs
@@ -6,13 +6,13 @@ namespace Military_Elite2.Models
 {
     public class Spy : Soldier
     {
-        public Spy(int id, string firstName, string lastName, int codeNumber) 
+        public Spy(int id, string firstName, string lastName, decimal codeNumber) 
             : base(id, firstName, lastName)
         {
             CodeNumber = codeNumber;
         }
 
-        public int CodeNumber { get; private set; }
+        public decimal CodeNumber { get; private set; }
         public override string ToString()
         {
             return base.ToString() + $"Code Number: {CodeNumber}";
